Reworked some movement things and the button mapping

+++ engine/src/IO/ButtonInfo.h
@@ -17,7 +17,7 @@ namespace engine {
     namespace IO {
         struct ButtonInfo{
             int buttonID, deviceID;
-            double xAxis, yAxis;
+            float xAxis, yAxis;
         };
     }
 }
